+++ src/config.h
@@ -1,0 +1,27 @@
+#ifndef CONFIG_H_
+#define CONFIG_H_
+
+/*************************
+ * MACROS
+ *************************/
+#define WIDTH         800
+#define HEIGHT        600
+#define BALL_SIZE     10
+#define BALL_SPEED    240
+#define PLAYER_WIDTH  20
+#define PLAYER_HEIGHT 75
+#define PLAYER_MARGIN 10
+#define PLAYER_SPEED  400.0f
+
+// https://www.wfonts.com/font/8bit-wonder
+/* #define FONT_NAME "8-BIT_WONDER.ttf" */
+// https://www.fontspace.com/pixeboy-font-f43730
+#define FONT_NAME "Pixeboy.ttf"
+
+/******** Colors *********/
+#define BLACK      0, 0, 0, 0
+#define WHITE      255, 255, 255, 255
+#define BLUE       0, 0, 255, 255
+#define RED        255, 0, 0, 255
+
+#endif // CONFIG_H_